pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sprint-goals-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "langchain>=0.2.14",
  "langgraph>=0.2.32",
  "langchain-core>=0.2.30",
  "langchain-openai>=0.1.20",
  "python-dotenv>=1.0.1",
  "typer>=0.12.3",
]

[project.scripts]
sprint-goals-agent = "agent.cli:app"

[tool.setuptools.packages.find]
include = ["agent*"]
exclude = ["reports*", "scripts*"]
